move validation tests to their own file
